integrate Login / Register

+++ src/Controller/DeckController.php
@@ -4,23 +4,28 @@ namespace App\Controller;
 
 use App\Entity\Deck;
 use App\Form\DeckType;
+use App\Security\Voter\DeckVoter;
 use App\Service\DeckService;
 use Symfony\Bridge\Doctrine\Attribute\MapEntity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
+#[IsGranted('ROLE_USER')]
 final class DeckController extends AbstractController
 {
     #[Route('/deck/list', name: 'deck_list', methods: ['GET'])]
-    public function list(DeckService $deckService): Response {
+    public function list(DeckService $deckService, Security $security): Response {
         $deck = new Deck();
+        $deck->setOwner($this->getUser());
 
         $form = $this->createForm(DeckType::class, $deck);
 
         return $this->render('deck/list.html.twig', [
-            'decks'=>$deckService->findAll(),
+            'decks'=>$deckService->findAllByOwner($security->getUser()),
             'form' => $form,
         ]);
     }
@@ -28,6 +33,7 @@ final class DeckController extends AbstractController
     #[Route('/deck/create', name: 'deck_create', methods: ['GET', 'POST'])]
     public function create(Request $request, DeckService $deckService): Response {
         $deck = new Deck();
+        $deck->setOwner($this->getUser());
 
         $form = $this->createForm(DeckType::class, $deck);
         $form->handleRequest($request);
@@ -54,6 +60,8 @@ final class DeckController extends AbstractController
             throw new \LogicException('This deck cannot be edited because it is used as a default deck.');
         }
 
+        $this->denyAccessUnlessGranted(DeckVoter::DELETE, $deck);
+
         $form = $this->createForm(DeckType::class, $deck);
         $form->handleRequest($request);
 
@@ -75,6 +83,8 @@ final class DeckController extends AbstractController
         if ($deck->isSystem()) {
             throw new \LogicException('This deck cannot be deleted because it is used as a default deck.');
         }
+
+        $this->denyAccessUnlessGranted(DeckVoter::DELETE, $deck);
 
         $deckService->deleteDeck($deck);
